Changes to the Team and Match services and controllers
Added update methods to the Match and Team services and implemented them in the Controllers also added a check when updating to see if the URL ID and the RequestBody ID match

+++ Java/Java 2024 Final Project Assignment/football-api/src/main/java/com/sirma/footballapi/service/TeamService.java
@@ -2,13 +2,16 @@ package com.sirma.footballapi.service;
 
 import com.sirma.footballapi.models.Team;
 import com.sirma.footballapi.repository.TeamRepository;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.annotation.Validated;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@Validated
 public class TeamService {
 
     @Autowired
@@ -22,11 +25,24 @@ public class TeamService {
         return teamRepository.findById(id);
     }
 
-    public Team createTeam(Team team){
+    public Team createTeam(@Valid Team team){
+        if(teamRepository.existsById(team.getId())){
+            throw new IllegalArgumentException(STR."Team with ID: \{team.getId()} allready exist");
+        }
+        return teamRepository.save(team);
+    }
+
+    public Team updateTeam(Long id, @Valid Team team){
+        if(!teamRepository.existsById(id)){
+            throw new IllegalArgumentException(STR."Team with ID: \{id} doesn't exist");
+        }
         return teamRepository.save(team);
     }
 
     public void deleteTeam(Long id){
+        if(!teamRepository.existsById(id)){
+            throw new IllegalArgumentException(STR."Team with ID: \{id} doesn't exist");
+        }
         teamRepository.deleteById(id);
     }
 }
